feat(grade): enhance web and mobile layouts with new summary row and table components, and improve score formatting.
